requirements.txt
psycopg2-binary
pandas
polars
pyarrow
scikit-learn
joblib
fastapi
uvicorn[standard]
python-dotenv
lightgbm